feat: Add server import in queries.ts
This commit adds a server import statement in the queries.ts file. This import is necessary for the file to work properly.

The commit also removes a console.log statement from the getInitials function in strings.ts. This console.log statement was no longer needed and has been removed.

Lastly, the commit updates the page.tsx file. It replaces the import statement for the TeamMembersCard component with the import statement for the new TeamMembers component. It also adds a new import statement for the TeamCars component. The code inside the return statement is updated accordingly to use the new components.